exception delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.educandoweb.course.services;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
-import com.educandoweb.course.services.exceptions.ResouceNotFoundException;
+import com.educandoweb.course.services.exceptions.DataBaseException;
+import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,7 +31,7 @@ public class UserService {
     public User findById(Long id){
         Optional<User> obj = repository.findById(id);
         //implementa a exceção criada
-        return obj.orElseThrow(()-> new ResouceNotFoundException(id));
+        return obj.orElseThrow(()-> new ResourceNotFoundException(id));
     }
 
     public User insert(User user){
@@ -36,7 +39,13 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DataBaseException(e.getMessage());
+        }
     }
 
     public User update(long id, User obj){
